Fixed SPI LASTXFER flag usage
The flag is the 24th bit in the register so applying it to the 16bit
data doesn’t make much sense. The CS value gets applied to the
transmit register in SPI_Write though, so LASTXFER can be used with
that (and NPCS is a 32bit variable in SPI_Write).