SPI: using CPOL instead of NCPHA works the same
NCPHA is the inverse of well-known CPHA,
so this means I’m using CPOL=CPHA=1 now, that’s the mode I think
I’m seeing in the datasheet of AD7793.

Previously with NCPHA=1 and CPOL=0, i.e. CPHA=CPOL=0, it worked
the same way, so it seems.

And there are still the same problems when reading STAT or CONF.